+++ packages/auth-construct/src/construct.ts
@@ -28,6 +28,7 @@ import {
 import { FederatedPrincipal, Role } from 'aws-cdk-lib/aws-iam';
 import { AuthOutput, authOutputKey } from '@aws-amplify/backend-output-schemas';
 import {
+  AttributeMapping,
   AuthProps,
   EmailLoginSettings,
   ExternalProviderOptions,
@@ -426,7 +427,7 @@ export class AmplifyAuth
       emailSettings.verificationEmailStyle !== 'LINK'
     ) {
       emailBody = emailSettings.verificationEmailBody(
-        VERIFICATION_EMAIL_PLACEHOLDERS.CODE
+        () => VERIFICATION_EMAIL_PLACEHOLDERS.CODE
       );
       if (!emailBody.includes(VERIFICATION_EMAIL_PLACEHOLDERS.CODE)) {
         throw Error(
@@ -438,10 +439,14 @@ export class AmplifyAuth
       emailSettings.verificationEmailBody &&
       emailSettings.verificationEmailStyle === 'LINK'
     ) {
-      emailBody = emailSettings.verificationEmailBody(
-        VERIFICATION_EMAIL_PLACEHOLDERS.LINK
-      );
-      if (!emailBody.includes(VERIFICATION_EMAIL_PLACEHOLDERS.LINK)) {
+      let linkText: string = '';
+      emailBody = emailSettings.verificationEmailBody((text?: string) => {
+        linkText = text
+          ? `{##${text}##}`
+          : VERIFICATION_EMAIL_PLACEHOLDERS.LINK;
+        return linkText;
+      });
+      if (linkText === '' || !emailBody.includes(linkText)) {
         throw Error(
           "Invalid email settings. Property 'verificationEmailBody' must utilize the 'link' parameter at least once as a placeholder for the verification link."
         );
@@ -601,7 +606,9 @@ export class AmplifyAuth
                   email: ProviderAttribute.GOOGLE_EMAIL,
                 }
               : undefined),
-            ...googleProps.attributeMapping,
+            ...this.convertToCognitoAttributeMapping(
+              googleProps.attributeMapping
+            ),
           },
           scopes: googleProps.scopes,
         }
@@ -622,7 +629,9 @@ export class AmplifyAuth
                   email: ProviderAttribute.FACEBOOK_EMAIL,
                 }
               : undefined),
-            ...external.facebook.attributeMapping,
+            ...this.convertToCognitoAttributeMapping(
+              external.facebook.attributeMapping
+            ),
           },
         }
       );
@@ -643,7 +652,9 @@ export class AmplifyAuth
                   email: ProviderAttribute.AMAZON_EMAIL,
                 }
               : undefined),
-            ...external.loginWithAmazon.attributeMapping,
+            ...this.convertToCognitoAttributeMapping(
+              external.loginWithAmazon.attributeMapping
+            ),
           },
         }
       );
@@ -664,7 +675,9 @@ export class AmplifyAuth
                   email: ProviderAttribute.APPLE_EMAIL,
                 }
               : undefined),
-            ...external.signInWithApple.attributeMapping,
+            ...this.convertToCognitoAttributeMapping(
+              external.signInWithApple.attributeMapping
+            ),
           },
         }
       );
@@ -703,7 +716,9 @@ export class AmplifyAuth
                     },
                   }
                 : undefined),
-              ...provider.attributeMapping,
+              ...this.convertToCognitoAttributeMapping(
+                provider.attributeMapping
+              ),
             },
           }
         );
@@ -718,7 +733,9 @@ export class AmplifyAuth
         `${this.name}SamlIDP`,
         {
           userPool,
-          attributeMapping: saml.attributeMapping,
+          attributeMapping: this.convertToCognitoAttributeMapping(
+            saml.attributeMapping
+          ),
           identifiers: saml.identifiers,
           idpSignout: saml.idpSignout,
           metadata: {
@@ -760,6 +777,43 @@ export class AmplifyAuth
     return result;
   };
 
+  /**
+   * Converts the simplified mapping type to cognito.AttributeMapping.
+   * @param mapping the AttributeMapping to convert to a cognito.AttributeMapping
+   * @returns cognito.AttributeMapping
+   */
+  private convertToCognitoAttributeMapping = (
+    mapping?: AttributeMapping
+  ): cognito.AttributeMapping | undefined => {
+    if (!mapping) {
+      return undefined;
+    }
+    const result: Record<
+      string,
+      | ProviderAttribute
+      | {
+          [key: string]: ProviderAttribute;
+        }
+    > = {};
+    for (const [attrName, value] of Object.entries(mapping)) {
+      if (typeof value === 'string') {
+        result[attrName] = {
+          attributeName: value,
+        };
+      }
+      if (typeof value === 'object' && attrName === 'custom') {
+        // dealing with custom attributes
+        const customAttributes: Record<string, ProviderAttribute> = {};
+        for (const [customKey, attrName] of Object.entries(value)) {
+          customAttributes[customKey] = {
+            attributeName: attrName,
+          };
+        }
+        result[attrName] = customAttributes;
+      }
+    }
+    return result;
+  };
   /**
    * Convert scopes from string list to OAuthScopes.
    * @param scopes - scope list
@@ -810,7 +864,7 @@ export class AmplifyAuth
           if (treatedAttributeName) {
             return [
               ...acc,
-              treatedAttributeName.userpoolAttributeName.toUpperCase(),
+              treatedAttributeName.userpoolAttributeName.toLowerCase(),
             ];
           }
         }
@@ -822,16 +876,16 @@ export class AmplifyAuth
     if (this.computedUserPoolProps.signInAliases) {
       const usernameAttributes = [];
       if (this.computedUserPoolProps.signInAliases.email) {
-        usernameAttributes.push('EMAIL');
+        usernameAttributes.push('email');
       }
       if (this.computedUserPoolProps.signInAliases.phone) {
-        usernameAttributes.push('PHONE_NUMBER');
+        usernameAttributes.push('phone_number');
       }
       if (
         this.computedUserPoolProps.signInAliases.preferredUsername ||
         this.computedUserPoolProps.signInAliases.username
       ) {
-        usernameAttributes.push('PREFERRED_USERNAME');
+        usernameAttributes.push('preferred_username');
       }
       if (usernameAttributes.length > 0) {
         output.usernameAttributes = JSON.stringify(usernameAttributes);
@@ -841,10 +895,10 @@ export class AmplifyAuth
     if (this.computedUserPoolProps.autoVerify) {
       const verificationMechanisms = [];
       if (this.computedUserPoolProps.autoVerify.email) {
-        verificationMechanisms.push('EMAIL');
+        verificationMechanisms.push('email');
       }
       if (this.computedUserPoolProps.autoVerify.phone) {
-        verificationMechanisms.push('PHONE');
+        verificationMechanisms.push('phone_number');
       }
       if (verificationMechanisms.length > 0) {
         output.verificationMechanisms = JSON.stringify(verificationMechanisms);