+++ packages/auth-construct/src/index.ts
@@ -9,10 +9,14 @@ export {
   OidcProviderProps,
   SamlProviderProps,
   ExternalProviderOptions,
+  VerificationEmailWithCode,
+  VerificationEmailWithLink,
   MFA,
   MFASettings,
   PhoneNumberLogin,
   TriggerEvent,
+  IdentityProviderProps,
+  AttributeMapping,
 } from './types.js';
 export { AmplifyAuth } from './construct.js';
 export { triggerEvents } from './trigger_events.js';