+++ packages/cli/src/command_middleware.test.ts
@@ -103,7 +103,7 @@ void describe('commandMiddleware', () => {
         } catch (err) {
           assert.match(
             (err as Error).message,
-            /Failed to load default aws region/
+            /Failed to load default AWS region/
           );
         }
       });
@@ -117,7 +117,7 @@ void describe('commandMiddleware', () => {
         } catch (err) {
           assert.match(
             (err as Error).message,
-            /Failed to load aws credentials for profile/
+            /Failed to load AWS credentials for profile/
           );
         }
       });
@@ -161,7 +161,7 @@ void describe('commandMiddleware', () => {
         } catch (err) {
           assert.match(
             (err as Error).message,
-            /Failed to load default aws credentials/
+            /Failed to load default AWS credentials/
           );
         }
       });
@@ -177,7 +177,7 @@ void describe('commandMiddleware', () => {
         } catch (err) {
           assert.match(
             (err as Error).message,
-            /Failed to load default aws region/
+            /Failed to load default AWS region/
           );
         }
       });
@@ -202,7 +202,7 @@ void describe('commandMiddleware', () => {
         } catch (err) {
           assert.match(
             (err as Error).message,
-            /Failed to load aws credentials for profile/
+            /Failed to load AWS credentials for profile/
           );
         }
       });
@@ -218,7 +218,7 @@ void describe('commandMiddleware', () => {
         } catch (err) {
           assert.match(
             (err as Error).message,
-            /Failed to load aws region for profile/
+            /Failed to load AWS region for profile/
           );
         }
       });