+++ packages/client-config/src/client-config-writer/client_config_writer.ts
@@ -5,6 +5,7 @@ import {
   ClientConfigFileBaseName,
   ClientConfigFormat,
   ClientConfigVersion,
+  GenerateClientConfigToFileResult,
 } from '../client-config-types/client_config.js';
 import { ClientConfigFormatter } from './client_config_formatter.js';
 
@@ -38,10 +39,8 @@ export class ClientConfigWriter {
     clientConfig: ClientConfig,
     version: ClientConfigVersion,
     outDir?: string,
-    format: ClientConfigFormat = ClientConfigFormat.JSON,
-    // TODO: update this type when Printer interface gets defined in platform-core.
-    log?: (message: string) => void
-  ): Promise<void> => {
+    format: ClientConfigFormat = ClientConfigFormat.JSON
+  ): Promise<GenerateClientConfigToFileResult> => {
     const targetPath = await this.pathResolver(
       this.nameResolver(version),
       outDir,
@@ -49,6 +48,9 @@ export class ClientConfigWriter {
     );
     const fileContent = this.formatter.format(clientConfig, format);
     await this.fsp.writeFile(targetPath, fileContent);
-    log?.(`File written: ${path.relative(process.cwd(), targetPath)}`);
+
+    return {
+      filesWritten: [path.relative(process.cwd(), targetPath)],
+    };
   };
 }